requirements.txt
asgiref==3.5.2
Django==4.1.3
django-colorfield==0.7.2
django-filter==22.1
django-phone-field==1.8.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
Pillow==9.3.0
pytz==2022.6
PyYAML==6.0
sqlparse==0.4.3
tzdata==2022.5
